add info attribute to MatchedParam
This new attribute adds the possibility to register extra information about a context parameter value.
For instance, with date-time entities, the info attribute can contain a Map with an entry for each part of the date-time (year, month, day...) indicating if it was explicitly said by the user (true value) or not (false value). The used NER system could fill the missing parts of 'April 4th', (year, hour, minutes and seconds) with the current date-time. But checking the info attribute, the bot developer could know which parts were not present initially (maybe to reconfirm them asking again to the user)